requirements.txt

python-dotenv
google_generativeai
flask
openai==0.28
groq